01-Medical_vision_app/requirements.txt
streamlit
python-dotenv
google-generativeai
Pillow